backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
sqlalchemy[asyncio]==2.0.35
asyncpg==0.29.0
alembic==1.13.0
pydantic==2.8.0
pydantic-settings==2.4.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pyotp==2.9.0
cryptography==43.0.0
python-multipart==0.0.9
httpx==0.27.0
redis==5.0.8
celery==5.4.0
slack-sdk==3.31.0
apscheduler==3.10.4
pykrx==1.0.46
pandas==2.2.2
pandas-ta
numpy==1.26.4
vectorbt==0.26.2
python-dotenv==1.0.1
qrcode[pil]==7.4.2
